rename : region class 모듈의 독립성을 위한 파일 위치 변경
