+++ src/main/java/ddd/caffeine/ratrip/module/place/model/Address.java
@@ -6,7 +6,7 @@ import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.validation.constraints.NotNull;
 
-import ddd.caffeine.ratrip.module.Region;
+import ddd.caffeine.ratrip.common.model.Region;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
